adding basic animal card and habitat themes
I added to the html and css files so they now include a basic animal card, which I've named #completedCard.  I've also included (and commented out) the habitat theme colors at the bottom of styles.css.  The themes can be added to the #completedCard and .animalDetails h3 class for color.